edit 1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -3,6 +3,7 @@
  * Auth: Mohammad Abulkhair
  */
 
+#include <stdio.h>
 #include "function_pointers.h"
 
 /**
@@ -17,7 +18,7 @@
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
     if (array && size && action) {
-        unsigned int i;
+        size_t i;
         for (i = 0; i < size; ++i)
         {
             action(array[i]);
